requirements.txt
python-whois
mysql-connector-python
requests